rename 'News' types for 'models'
