+++ pages/contacto.tsx
@@ -9,7 +9,7 @@ import RrssIcon from 'components/RrssIcon/RrssIcon';
 import ContactItem from 'components/ContactItem/ContactItem';
 import { BG_CONSTANTS, MAPS_URL } from 'utils/constants';
 import { GetStaticProps } from 'next';
-import { General } from 'types/News';
+import { General } from 'types/models';
 
 type Phones = General & {
   title: General['title'];
